Make the SSA instruction bs_put conform to failure conventions
There is a convention in SSA code that an instruction that can fail
must be immediately followed by a `succeeded:Kind` instruction. For
example, a BIF call in a body would look like this:

    Result = bif:Bif Arg, ...
    Bool = succeeded:body Result
    br Bool, ^Success, ^Fail

Because the `bs_put` instruction does not have any return value,
we did not originally follow the failure convention, but instead
made it behave like a guard test:

    Bool = bs_put Type, ...
    br Bool, ^Success, ^Fail

That saves an instruction, but it necessitates special handling
of `bs_put` in some optimizations. To avoid that special handling,
change the representation of `bs_put` to (in a body):

    Result = bs_put Type, ...
    Bool = succeeded:body Result
    br Bool, ^Success, ^Fail

Or (in a guard):

    Result = bs_put Type, ...
    Bool = succeeded:guard Result
    br Bool, ^Success, ^Fail

`Result` is here a dummy variable that will be discarded in the
translation to BEAM code.