syntax_tools: Fix merl handling of erl_anno:anno()
Merl currently assumes all annotations are locations, which can be
either an integer or a two-tuple of integers. However, erl_anno supports
more fields, like file and text. This is done through an opaque type,
which is actually a proplist.

OTP adds support for column numbers, and it would be nice if merl would
support those somewhere down the line. A first step would be to allow
the quote function(s) to accept the full erl_anno:anno() type. It
currently pulls the location and then continues as normal.

I also renamed Line to Location in merl_transform to make it clear that
it can be both a integer and tuple of integers. Later perhaps a full
anno, but not now.