Speed up compilation of huge modules
Reorder the compiler passes so that the beam_ssa_share passes
is run before the beam_ssa_bool and beam_ssa_recv passes.
For huge modules where the share optimization can be applied,
that can noticeably reduce the compile time.

For example, on my computer, this change reduced the compile
time for the String.Unicode module in the Elixir standard library
from almost 30 seconds down to less than 15 seconds.