ssl: Make sure TLS distribution connection processes has a common session table
Use monitor for the listen socket instead of a monitor for the process that opened
the listen socket for all connection processes, this is a better solution as
ownership of the socket can be transfered and listen socket could be closed
without the process that opened it terminating.

Provide backwards compatible solution for TCP upgrade TLS servers
We like to later introduce an option for them to to provide a listen
socket so that the server instance may be properly identified and handled.
For now we keep at most one session table that will be Erlang node global
for TCP upgrade TLS servers, such as the Erlang distribution servers.