jit: Emit i_select_val_bins as code.
The i_select_val_bins instruction has previously been implemented by a
call to a global subroutine which does a binary search in a
table. This patch changes the code generator to emit the binary search
as code.

The changed code generation strategy provides varying speedups
compared to the table search, depending on the number of choices and
size of the values searched for, but is in no case slower than the
legacy implementation.

For i_select_val_bins-instructions with more than 11 and less than 33
choices, a speedup of more than 40% can be expected. For a larger
number of choices (<= 4096), but where the largest key's tagged form
fits in a 16 bit immediate, a speedup of between 40% to 65% can be
expected. If the key values are too large to fit in a 16 bit
immediate, performance falls off when the table gets larger. Peak
performance at a speedup of more than 60% for small keys (tagged value
fits in 16 bits) occurs at a table size of 512 elements, and falls off
to 40% when the table contains 4096 keys. For large keys (the tagged
value does not fit in 32 bits), peak performance occurs at a table
size of 256 choices with a speedup of 70%, but performance falls off
to 25% when the number of choices is increased to 4096.

For small keys (tagged small integer fits in a 16 bit word), memory
use is reduced by around 5%. With large keys (the tagged value does
not fit in 32 bits), the size required for a i_select_val_bins
instruction can increase by up to 10%.