Reduce confusion by rephrasing some compiler warnings
Consider this module:

    -module(w).
    -export([ll/0, sum/0, tt/0, ignored/2]).

    ll() ->
        L = {a,b,c},
        length(L).

    sum() ->
        A = a,
        B = b,
        A + B.

    tt() ->
        (try nan after whatever end) + 42.

    ignored(X, Y) ->
        X + Y,
        ok.

When compiling this module using https://github.com/erlang/otp/pull/3020,
the following warnings are emitted:

    w.erl:6:5: Warning: this expression will fail with a 'badarg' exception
    %    6|     length(L).
    %     |     ^

    w.erl:11:7: Warning: this expression will fail with a 'badarith' exception
    %   11|     A + B.
    %     |       ^

    w.erl:14:34: Warning: this expression will fail with a 'badarith' exception
    %   14|     (try nan after whatever end) + 42.
    %     |                                  ^

    w.erl:17:7: Warning: the result of the expression is ignored (suppress the warning by assigning the expression to the _ variable)
    %   17|     X + Y,
    %     |       ^

The last three warnings are slightly confusing, because the warnings
say "this expression" or "the expression", but the column position
points out the `+` operator in the middle of the expresson instead
of the beginning of the expression.

For the second warning, it would be better to point out the variable `A`:

    w.erl:11:5: Warning: this expression will fail with a 'badarith' exception
    %   11|     A + B.
    %     |     ^

However, to produce that warning would need a major rethinking of the
compiler pass, `sys_core_fold`, that produces that kind of warning.
`sys_core_fold` is an optimization pass that opportunistically emits
warnings while optimizing the code. When the warning for line 11 is
emitted, the original code has been lost and the line and column
number for the variable `A` is no longer available.

The third warning does not seem possible to fix by moving the position of the caret:

    w.erl:14:34: Warning: this expression will fail with a 'badarith' exception
    %   14|     (try nan after whatever end) + 42.
    %     |                                  ^

Pointing to the atom `nan` would be confusing.

Considering that the second warning would need a lot of work and that there is no good
solution for the third warning, I instead chose to rephrase the warnings like this:

    w.erl:6:5: Warning: the call to length/1 will fail with a 'badarg' exception
    %    6|     length(L).
    %     |     ^

    w.erl:11:7: Warning: evaluation of operator '+'/2 will fail with a 'badarith' exception
    %   11|     A + B.
    %     |       ^

    w.erl:14:34: Warning: evaluation of operator '+'/2 will fail with a 'badarith' exception
    %   14|     (try nan after whatever end) + 42.
    %     |                                  ^

    w.erl:17:7: Warning: the result of evaluating operator '+'/2 is ignored (suppress the warning by assigning the expression to the _ variable)
    %   17|     X + Y,
    %     |       ^