dialyzer: Add option 'error_location'
Since PR 3006 column numbers are by default included in abstract code,
which implies that Dialyzer includes column numbers in warnings. This
can be a problem for tools using Dialyzer, such as Rebar3.

It does not suffice to say that all analyzed files are to be compiled
with `{error_location, line}' (can be set in ERL_COMPILER_OPTIONS)
since Dialyzer can also analyze source code.

Therefore, the `error_location' option is introduced in Dialyzer as
well. The environment variable ERL_COMPILER_OPTIONS is also
recognized.