[kernel|gen-tcp-socket] Adjust test case and doc for async shutdown
The behaviour of the inet driver with regard to send and shutdown
is not matched if the inet_backend = socket.

If the gen_tcp:send function is called with more data than the OS-buffers
can hold, the inet driver will "send" what it can and store the rest
in internal buffers, then select and report back ok to the caller.
The data in the buffers, will be sent by the inet driver asynch, without
any involvement of the sender process.

The the gen_tcp:shutdown(Socket, write | read_write) is called, while
this is in effect, the inet driver will simply make a note of this.
Then, when the data in the buffers has been sent, the shutdown operation
will be performed.

In the first case, the behaviour will instead be that the send function
will hang until the data has been fully sent, or the send timeout ends.
In the second case, the shutdown operation will be done immediately,
thereby interrupting any ongoing write operations.

OTP-17156