Remove old mac init code
wxWidgets does it in its init code and it causes harm on newer MacOs's,
i.e. the top-menus doesn't work.

A fix is merged to wxWidgets master branch and will be released as
wxWidgets-3.1.5 until then top menues will not work on mac.

On old macos the old init code in wxWidgets still work and is not
needed in wx driver init.