beam_block: Remove uncovered handling of make_fun3 in is_unused/2
The clause for handling make_fun3 does not seems possible to cover;
because make_fun3 takes a list of registers in any order, there will
never be any register swapping instructions before it.

In any case, this change is perfectly safe because of the default
clause returning false that will simply prevent the replacing of the
three move instructions with a swap instruction.