ssl: Avoid updating server session cache API
TLS session cache (pre TLS-1.3 that has a different session mechanism)
was rewritten inspired by the new mechanism to have a hard max value and still
have as good reuse rate as possible. This was patched to solve a real problem
at the server side, with some API and client side enhancements remaining.

After enhancing the client, rewrite server solution so to avoid
changing callback API, as it aims to make it possible to have
a session table that survives a node restart and hence assumes a data
base reference that is not updated. As this is a legacy mechanism