Fully support IEEE-754 floats on binary matching
At the moment, Erlang does not fully support IEEE-754
encoding in binary matching. In particular, it is not
possible to decode/encode infinity, negative infinity,
and nan's.

Furthermore, all of the arithmetic operations in Erlang
raise in the presence of the values above.

The IEEE-754 standard makes a strong case for *not*
raising when working with non-finite values - but we can
probably say this ship has sailed for Erlang. Furthermore,
one can assume that, if those features were desired by
the community, they would already have been implemented.

Therefore, this patch proposes to allow encoding/decoding
of non-finite values. After all, if a non-finite value was
encoded as part of a binary, it was likely expected that
the decoding party should handle it. This pull request adds
basic mechanisms to do so.

At the same time, this pull request does not change
(nor intends to) any of the arithmetic operations. Therefore,
operations that raised in the past when returning non-finite
values continue to behave the same.

A possible venue for exploration in a later pull request
is to augment the math module with IEEE compatible functions,
for example, `math:ieee_log/1`, `math:ieee_exp/1`, as well as
`math:ieee_add/2`, `math:ieee_subtract/2` and so forth
(including `math:ieee_negate/1`) that supports non-finite
values. This would allow Erlang developers to work with
non-finite values only if they want to and allow them to
do so efficiently (as currently it is inneficient and
cumbersome).

I am submitting this pull request for a initial request for
comments. In particular, I want to show supporting non-finite
values seems to be trivial, given this is handled by any C
toolchain that supports IEEE-754 (which is commonplace).
I will work on tests if there is interest.