sasl: Fix no_dot_erlang testcase
Set HOME to priv_dir in order to not have to change
the ~/.erlang cookie on the machine we do the test on.
We do not want to change the ~/.erlang as there could
be other erlang services running that would be unhappy
when it is changed. For instance the erlang compiler.