kernel: Let option 'size' of disk_log:open/1 change size
If the value of option 'size' does not match existing wrap log size,
the 'size_mismatch' error is no longer returned, but the size of the
log changed. This applies only if the disk log is not already open.

We have made this incompatible modification to help customers handle
upgrade scenarios.

See also ERL-1418, GH-4469 and ERIERL-537.