Add test.
With this change to the test but not the fix in the previous commit, running
os_signal_SUITE with emulator arg +Bd or +Bi will fail (BEAM will die with a
signal, which is not the best way to fail a test, but we don't have a shell
test suite to catch it).

To confirm, you can use ts, with
`ts:run(emulator, os_signal_SUITE, [{vars, [{erl_start_args, "+Bd"}]}]).`.