Ensure that Xref does not see Debugger as a dependency for Kernel
The `error_handler` module in the Kernel application has a call to the
`int` module in the Debugger application. That call is only there to
make the Debugger work. To avoid that Xref sees Debugger as a
dependency for Kernel, the call to the `int` module is obfuscated in
the following way:

    breakpoint(Module, Func, Args) ->
        (int()):eval(Module, Func, Args).
    int() -> int.

That is, the call will show up as a call to an unknown module in
Xref.

With the improved whole-module type analysis introduced in
294d66a295f6 (#2100), the compiler sees through the obfuscation
and rewrites the code to:

    breakpoint(Module, Func, Args) ->
        int:eval(Module, Func, Args).

It would be fun to solve this issue by introducing further obfuscations,
but such tricks could stop working in the future when the compiler is
improved. Instead, turn off whole-module type analysis by compiling
the module with the `no_module_opt` option.

Resolves #4546