otp_SUITE: Enhance the dependency tests to run in both xref modes
Xref can be run in two modes. More detailed information can be
found by running in `functions` mode, but it requires analysing
source code or BEAM files compiled with the `debug_info` option.
When no debug_info is available, xref must be run in `module`
mode.

Note that more module edges can potentially be found in `modules` mode
because the analysis is based on the BEAM code after all
optimizations. For example, an apply in the source code could after
optimizations be resolved to a specific function.

Update otp_SUITE to test runtime dependencies in both modes.
It is important to test the `modules` modes because reltool runs xref
in that mode (because the BEAM files to be released might not contain
debug information).