Enhance the optimization of the try/catch elimination optimization
The `beam_ssa_opt:ssa_opt_try/1` optimization would sometimes fail to
eliminate a redundant `try`/`catch` construct because it only looked
for `new_try_tag` instructions at the beginning of an SSA block. Here
is an example:

    foobar({Anno}=T, V) ->
        try
            setelement(1, T, V)
        catch
            _Class:_Error:_Stack ->
                Anno
        end.

In the SSA code, the `get_tuple_element` instruction from the tuple
matching in the function head is at the beginning of the block with the
`new_try_tag` instruction:

    8:
      _10 = get_tuple_element _0, `0`
      @ssa_catch_tag = new_try_tag `'try'`
      br @ssa_catch_tag, ^6, ^5

Generalize the optimization to handle this example. While at it,
also make the optimization run faster by eliminating redundant
set update operations.