Fix multi-zero timeout callback handling in timer wheel
Execution of a zero timeout callback which was started from another
zero timeout callback could be delayed. Most timers in the system
do not behave like this. The service timer placed in the timer wheel
for management of the red-black tree of very long timers may however
behave like this.