[kernel|gen-tcp-socket] Make gen_tcp_socket bug-compatible with inet-driver
The inet-driver issues econnreset when a read fails (closed),
regardless of the actual error.
On some platforms actually this is the actual error, but this
is not universal. Now we make sure that if the user has
configured 'show_econnreset = true', a tcp_error message will
be sent with this reason when 'closed' is the result.

OTP-17156