beam_ssa_dead: Avoid unnecessary sorting of the switch list
One of the invariants maintained by beam_ssa_lint is that the
list of values and labels in a switch is sorted.

Therefore, beam_ssa_dead can safely assume that the list is sorted,
and can also avoid calling beam_ssa:normalize/1 after optimizations
of switches that only update labels.