Add encode function
This is one of the two big changes in this changeset.

I really want this function to exist so that I can take
a string and convert it to the URL-encoded form. This
probably isn't the best way to do this - it seems like
high-unicode characters are going to be caught by this -
but it's a place to start.