Refactor is_reserved/1
Poking around the encode/decode functions, I saw the
is_reserved/1 and is_sub_delim/1 functions, and I
thought that the logic seemed redundant. If a
codepoint is a sub_delim, it MUST be a reserved
codepoint. Similarly, if a codepoint is a gen_delim
as specified by RFC 3986, it must also be a reserved
codepoint. This seems to decompose very nicely. I
tried to extract the logic into a smaller function and
compose the is_sub_delim/1 function with the
is_gen_delim/1 function to create a shorter and
easier-to-alter is_reserved/1 function. Note the use
of orelse as a lazy-or in order to prevent wasting
CPU cycles.