[megaco|test] Tweaked load suite
Actually what was tweaked was the start of the utility mgc and mg
used by "all" megaco test suites.
Instead of using spawn_link, use spawn_monitor.
A problem detected in one of the load test cases eaddrinuse
was supposed to be catched (and converted to a skip) in the mgc
start (there was code for it) but was not. Instead it was catched
at the top level, as an error.