erlexec: Add checks that next argument is valid
The argument -- is silently added in a number of
places in order to seperate arguments from different
sources. When checking to see if an argument that
erlexec needs to handle is valid we therefore check
that it does not start with '--'.

Since `+S -1` is a valid argument we cannot make all
`-` and `+` values invalid, which is a shame but not
much we can do about that.

Closes #4624