[kernel|gen-tcp-socket] Partial fix for socket option pktoptions
This is a special case as this option is only valid on Linux
and then is basically unsupported. The RFC this option is based
on is RFC 2292 which has been obsoleted by RFC 3542 (which
no longer includes this option). This later RFC solves the
problem in another way.
Currently the 'socket' (and nif) do not actually implement this
option.

OTP-17156