[kernel|socket] Add "support" for unsupported obsoleted option 'pktoptions'
Add "support" for the unofficial option 'pktoptions'.
This option was specified in a now obsoleted RFC 2292 (obsoleted
by RFC 3542). It is an IPv6 option, but was *also* implemented
for IPv6 on *Linux*. So, this *may* work on some Linux distros...
This has been added for backward compatibility reasons.
Currently it is only partly processed as it is intended to used
by gen_tcp_socket, which discards the parts it does not need.
Do *not* use!

OTP-17280 (OTP-17156)