docgen: Allow internal docs to contain both modules and chapters
If an application includes internal documentation which refers to both
modules and text in chapters, the xsl will produce two menus, one for
the modules and one for the chapters, which the CSS then will render
on top of each other making a real mess.

This is an attempt to avoid the described problem by introducing a new
menu template which is used when the internal documentation contains
both modules and chapters.