compiler: Document BEAM SSA invariants
The BEAM SSA representation is well described type-wise, but some
aspects such as inter-instruction requirements and strategies for
creating unique names are not. This patch adds a document to the
internal documentation of the compiler describing the structure and
format of BEAM SSA in the hope that it helps first-time compiler
developers.