Add pkg-config support for erl_interface
pkg-config is a helper tool used when compiling applications and
libraries.

This commit introduces the file erl_ei.pc which will be installed
in the standard pkg-config directory.

Thus, using pkg-config like this:

    pkg-config --libs --cflags erl_ei

Returns the correct libs and cflags that can be used by build tools
(like autotools);

    -I/cross/sysroots/core2-32-vendor-linux/usr/lib/erlang/lib/erl_interface-4.0.2/include
    -L/cross/sysroots/core2-32-vendor-linux/usr/lib/erlang/lib/erl_interface-4.0.2/lib -lei

In order to enable the pkg-config installation a new configure option,
--enable-pkg-config, was introduced. The default is not install the
pkg-config file. And one could enable it using --enable-pkg-config
argument when calling configure script.