ssl: Make sure incomplete chain is a prefix of the new chain candidate
When trying to reconstruct an incomplete chain we must not
use a constructed chain that is a shorter incomplete chain than
the original chain. This might happen if we have a chain that we
will not be able to reconstruct and hence verify. Even though this
is not a problem from a verification point (will fail with unknown_ca anyway)
it could cause different behavior for a verify_fun.

Closes #4682