[zip] Add execute permission to zipped directories (fix GH-4687)
Before this commit all files and directories in a ZIP-archive created
by `zip:create` or `zip:zip` got the same permission 644. This made
the files inside dictionaries inaccessible (even for the owner). After
this commit all zipped dictionaries get permission 744 (adds
executable permission for the owner), which makes the dictionary
content accessible for the owner.