Fix ct_gen_conn connection registration race
Module ct_netconfc (and more) creates transport connections in callbacks
from the process started by ct_gen_conn:start/4. The process pid is
written to a table in ct_util, but inserting and deleting entries took
place in different processes: insert from the calling process, delete
from the started process. As a result, a quick error from the underlying
transport (ssh in the case of ct_netconfc) could result in the
ct_gen_conn process deleting its entry and exiting before the calling
process had written it, which orphaned the entry in the table and lead
to errors from subsequent calls to ct_gen_conn. In particular, start/4
returned the pid of the dead process.

Move insert into the started process to avoid the race.