Fix multiple ct_gen_conn:do_within_time/2 issues
The function assumed the caller is trapping exits and is linked to the
connection process, but had several problems:

- It left an EXIT message from the temporary process evaluating the
  function whose value is to be returned in the caller's message queue
  in the successful case before timeout.
- It could leave both a result message and an EXIT message in the
  caller's message queue after connection process exit since both
  can be sent before the temporary process is unlinked and killed.
- It didn't return at all if the fun being evaluated failed before
  timeout since receiving the EXIT message matched on the reason
  being 'killed'.
- It potentially reordered the caller's message queue by resending
  an EXIT message to self().

Do away with the second assumption and all the problems by using the
exit reason to communicate the result from the temporary process and
monitoring on the connection process.

There's one more problem however: the connection process pid retrieved
from the process dictionary is only put there after the user's init/3
callback returns it, but ct_telnet calls do_within_time/2 from this
callback (and others). The link to the connection process is also added
after init/3 makes it known. This should be fixed by not passing
function arguments through the process dictionary, but fixing this in
the tortured mechanics of ct_telnet (the only caller of do_within_time/2
in practice) isn't worth the effort in this commit. Simply work around
it by using self() as the connection pid in this case, which renders the
monitoring as ineffectual as an unset link. The consequence is just that
failure of a connection process won't be detected in this case, so
do_within_time/2 times out instead in the error case.