Fix multiple ct_gen_conn issues
Redo as a gen_server to simplify the mechanics (no need to reinvent that
wheel) and address the following problems:

- A call/3 reply from the process could be left in the caller's message
  queue after a timeout.
- Failure of a user's terminate callback caused a call result to not be
  sent.
- Unexpected messages accumulated in the message queue with option
  forward = false.
- Documentation and specs were out of sync with code. Redo the edoc specs
  as type specs.

Exit reasons have changed slightly, wrapping them in a shutdown tuple to
avoid gen_server logging expected exits as failures, but these are
mapped back by start/4 and call/3 since test code matches on the values.
(Code that monitors on the exit reason doesn't seem to care.)

do_within_time/2 (called from user callbacks in ct_telnet) is one source
of unexpected messages, which will be addressed in a subsequent commit.

Untabified for consistent indentation.