Rewrite to use bag table
Using a set table broke the desired behaviour to get multiple
addresses for a domain.  Revert to using a bag table instead,
but still do lookups on the ets table without going through
the server.

The cache is now a bag table containing #dns_rr{} records
with domain as key.  A lookup uses ets:match_object for domain,
class and type so the key is in the match expression.  This is
a rather fast lookup, but it has to check all objects with
the domain (key) for matching class and type.

It might be faster to store tuples with a composite key
{{Domain, Class, Type} = Key, #dns_rr{}} since then ets:lookup
for Key could be used instead.  Due to how bag tables are
implemented internally it is not certain that this would be
very much faster.

After a lookup for all matching records, they need to be updated
with a new cnt time value.  Since we are using a bag table this
has to be done by inserting a new record and deleting the old.
If we insert before delete we always have a record in the cache
so parallel lookups will not accidentaly find no record.

This might create duplicate records i.e records with the same
(domain, class, type) but different e.g cnt field since parallel
processes might insert different new records and then delete
the old.  Therefore we have to weed out duplicate records
when returning the result.

Try again to use a set table?
-----------------------------
Faster lookups would be achieved by storing
{{Domain, Class, Type} = Key, [#dns_rr{}]} tuples in a set table.

Updates of the cnt value could then be done by overwriting
the tuple with a new [#dns_rr{}] list.  Then parallel processes
might overwrite each others' changes, which might accedentally
delete a cache record written by a recent parallel lookup.
This might actually be acceptable since it only would cause
another cache miss.