Remove duplicated code
PRF is specific to pre-TLSv1.3, so it should not be on ssl_gen_statem.
Indeed, the exact same code can be found in tls_dtls_connection, but a
quick look at test coverage said that it wasn't being called to begin
with. I believe that this duplication was introduced in ff15b5748d2.