Add the hello module as a test case
When developing a JIT, or a new implementation of Erlang, or a
new compiler for Erlang, one usually writes multiple small test cases
to use during development. Usually all that testing code is discarded
afterwards.

To avoid having to come up with those tests from scratch the next time
such a task is attempted, this commit adds a `hello` module that tests
most constructs that an Erlang implemenation needs to handle. To ensure
that it keeps working, we will embed it into a test suite for the
ERTS application.