Prepare for literals in non-message signals
Previously references to literals could only exist in non-message
signals. We, however, at least need to allow literals in messages
passed using the alias-message signal, but also other non-message
signals, such as for example exit signals, should be able to contain
literals.

The most straightforward idea would be to handle all types of signals
that can contain literals. This however become quite complicated since
there are a lot of different signals with different layouts. Instead
of this we pass copy-literal-area signals to all processes after having
set the specific area as "copy literal area". All new signals being
created, after the area has been set, will copy the literals on creation
and will therefore not contain any references to the area being
released. When the copy-literal-area signal is received any references
to the literal area will either have been dropped or converted into
messages placed in the message queue. That is, at this point only the
message queue has to be inspected. By this we don't have to take care
of a lot of different signal types. Also when introducing new signal
types carrying literals we wont have to bother about this.

When the copy-literal-area signal is received, the message queue
is checked and any literals are copied into the messages. After that
the heap is scanned for references to the literal area. If references to
the area are found, a request for literal GC is sent as reply to the
literal-area-collector process; otherwise, a confirmation that the process
is clean is sent as reply. The copy-literal signal is prepared so it can be
reused when sending the reply message to the literal-area-collector
process. By this we significantly reduce the amount of allocations made
for the whole literal GC operation.

If a literal GC is needed, this will be scheduled as a system task at a
later time by the copy-literal-collector process when appropriate (using
the same approach as before).